update db stuff
