+++ src/lib/db/entities/GuestInformation.ts
@@ -1,7 +1,4 @@
-import { sequelize } from '$lib/db/db';
-import { DataTypes, Model, ModelStatic } from 'sequelize';
-
-type GuestInformationAttributes = {
+export type GuestInformationAttributes = {
 	id: string;
 	festivalEventId: string;
 	food: string;
@@ -9,22 +6,4 @@ type GuestInformationAttributes = {
 	numberOfOtherGuests: number;
 	coming: boolean;
 	comment: string;
-}
-
-export const GuestInformation: ModelStatic<Model<GuestInformationAttributes, any>> = sequelize.define('GuestInformation', {
-		id: {
-			type: DataTypes.STRING,
-			primaryKey: true,
-			allowNull: false
-		},
-		festivalEventId: {
-			type: DataTypes.STRING,
-			allowNull: false
-		},
-		food: { type: DataTypes.STRING },
-		drink: { type: DataTypes.STRING },
-		numberOfOtherGuests: { type: DataTypes.INTEGER },
-		coming: { type: DataTypes.BOOLEAN },
-		comment: { type: DataTypes.STRING }
-	}
-);
+};
