+++ src/lib/models/FrontendUser.ts
@@ -1,4 +1,3 @@
 import type { BaseUser } from '$lib/models/BaseUser';
 
-export interface FrontendUser extends BaseUser {
-}
+export interface FrontendUser extends BaseUser {}
